app/requirements.txt
Django==2.2.20
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-oauth-toolkit==1.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4